+++ datatypes/datatypes/Particle.h
@@ -21,8 +21,8 @@ public:
   Particle();
   Particle(int pdgid, double charge, TLorentzVector tlv, double status = 0);
   Particle(int pdgid, double charge);
-  Particle(long id, int pdgid, double charge, TLorentzVector tlv, double status = 0);
-  Particle(long id, int pdgid, double charge);
+  Particle(long long id, int pdgid, double charge, TLorentzVector tlv, double status = 0);
+  Particle(long long id, int pdgid, double charge);
   std::string stringDescription() const;              ///< String to describe the particle
   const TLorentzVector p4() const { return m_tlv; }   ///< 4-momentum, px, py, pz, E
   const TVector3 p3() const { return m_tlv.Vect(); }  ///< 3-momentum px, py, pz
@@ -37,10 +37,10 @@ public:
   bool status() const { return m_status; }                ///<status code, e.g. from generator. 1:stable.
   TVector3 startVertex() const { return m_startVertex; }  ///<start vertex (3d point)
   TVector3 endVertex() const { return m_endVertex; }      ///<end vertex (3d point)
-  long id() const { return m_uniqueId; }
+  long long id() const { return m_uniqueId; }
 
 protected:
-  long m_uniqueId;  // to be used by virtual classes
+  long long m_uniqueId;  // to be used by virtual classes
 private:
   TLorentzVector m_tlv;
   int m_particleId;
