+++ datatypes/src/PFParticle.cpp
@@ -19,7 +19,7 @@
 
 namespace papas {
   
-PFParticle::PFParticle(long uniqueid,const Track& track) : //TODO check what this is used for
+PFParticle::PFParticle(long long uniqueid, const Track& track) : //TODO check what this is used for
     PFParticle(uniqueid,
             211 * track.charge(),
                TLorentzVector(track.p3(), track.energy()),
@@ -27,7 +27,7 @@ PFParticle::PFParticle(long uniqueid,const Track& track) : //TODO check what thi
             track.path()->field()) {
 }
 
-  PFParticle::PFParticle(long uniqueid,int pdgid, TLorentzVector tlv, TVector3 vertex, double field) :
+  PFParticle::PFParticle(long long uniqueid, int pdgid, TLorentzVector tlv, TVector3 vertex, double field) :
   Particle(uniqueid, pdgid, ParticlePData::particleCharge(pdgid), tlv),
   m_vertex(vertex),
   m_isHelix(fabs(charge())>0.5)
